requirements.txt
streamlit
google-generativeai
fpdf
python-dotenv